requirements.txt
# Core Analytics
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Financial Libraries
yfinance>=0.2.0
pyportfolioopt>=1.5.0
quantlib>=1.30.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Database
psycopg2-binary>=2.9.0
redis>=4.6.0
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
hypothesis>=6.75.0
pytest-cov>=4.1.0

# Monitoring
prometheus-client>=0.17.0

# Development
python-dotenv>=1.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0

# Frontend (if needed)
# react and typescript will be handled separately 